requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
llama-cpp-python
pyyaml
numpy